Fifthcommit

+++ src/main/java/com/example/GraduationProject/RepositoryLayer/ProductRepository.java
@@ -3,6 +3,7 @@ package com.example.GraduationProject.RepositoryLayer;
 import com.example.GraduationProject.Business.Entity.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductRepository extends JpaRepository<Product, Integer> {
+public interface ProductRepository extends JpaRepository<Product, Long> {
+
 
 }
